requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.7.0
django-ipware==7.0.1
frozenlist==1.7.0
idna==3.10
multidict==6.5.0
propcache==0.3.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-ipware==3.0.0
requests==2.32.4
sqlparse==0.5.3
twilio==9.6.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
